separete all profile components
